# config/openapi.yaml
openapi: 3.0.0
info:
  title: Trade Data API
  version: 1.0.0
paths:
  /upload:
    post:
      summary: Upload a CSV file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Successfully uploaded CSV file
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request, please upload a valid CSV file
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /vwapcalculator:
    post:
      summary: Calculate VWAP (Volume Weighted Average Price)
      responses:
        '200':
          description: Successfully calculated VWAP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeResponse'
        '400':
          description: Bad request, please provide valid trade data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /trades:
    get:
      summary: Get all trades
      responses:
        '200':
          description: Successfully retrieved all trades
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                    currencyPair:
                      type: string
                    timeStamp:
                      type: string
                      format: date-time
                    tradeType:
                      type: string
                    tradeAmount:
                      type: number
                      format: double
                    volume:
                      type: number
                      format: double
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
components:
  schemas:
    TradeResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            NZD/GBP:
              type: array
              items:
                $ref: '#/components/schemas/TradeSession'
    TradeSession:
      type: object
      properties:
        session:
          type: string
          example: "Hour during: 9 to 10"
        vwap:
          type: number
          format: double
          example: 0.4737097428172133